[#173322606] move modal components and logic in a separated component
